refactor(tests): split srv checker suite by scenario
Replace the monolithic SRV checker tests with required, optional, and error-focused modules, add shared SRV test support helpers, and preserve existing SRV coverage while tightening assertions on stable result messages.
